Put Chunk into its own file in preparation for collision detection
